@@ -1,6 +1,6 @@
 module github.com/CuteReimu/hksplitmaker
 
-go 1.17
+go 1.20
 
 require (
 	github.com/dlclark/regexp2 v1.9.0
